// src/types/enums.ts
export enum Item {
  home = 'home',
  password = 'password',
  recipe = 'recipe',
  letter = 'letter',
  animal = 'animal',
  number = 'number',
  color = 'color',
  country = 'country',
  cartoon = 'cartoon',
}

export enum CartoonEnum {
  pokemon = 'pokemon',
  ginger = 'ginger',
  arnold = 'arnold',
}